+++ infobigan-torch/setup.py
@@ -1,0 +1,8 @@
+from setuptools import setup
+
+setup(name='infobigan_torch',
+      version='0.1',
+      description='Torch implementation of InfoBiGAN',
+      url='https://github.com/Riley16/InfoBiGANs/tree/master/infobigan-torch',
+      license='MIT',
+      packages=['infobigan_torch'])
